+++ truffle/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This changelog summarizes major changes between Truffle versions relevant to languages implementors building upon the Truffle framework. The main focus is on APIs exported by Truffle.
 
+## Version 1.0.0 RC10
+* Added support for setting current working directory for TruffleFiles.
+
 ## Version 1.0.0 RC9
 
 * Added support for setting the `ThreadGroup` and `stackSize` on truffle thread creation in `TruffleLanguage.Env.createThread`.